+++ package.json
@@ -14,6 +14,28 @@
     "marketplace"
   ],
   "main": "dist/golem-js.cjs.js",
+  "exports": {
+    ".": {
+      "require": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/golem-js.cjs.js"
+      },
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/golem-js.es.js"
+      }
+    },
+    "./experimental": {
+      "require": {
+        "types": "./dist/experimental.d.ts",
+        "default": "./dist/golem-js-experimental.cjs.js"
+      },
+      "import": {
+        "types": "./dist/experimental.d.ts",
+        "default": "./dist/golem-js-experimental.es.js"
+      }
+    }
+  },
   "browser": "dist/golem-js.min.js",
   "module": "dist/golem-js.es.js",
   "types": "dist/index.d.ts",
