+++ src/index.ts
@@ -9,8 +9,6 @@ export * from "./network";
 export * from "./events";
 export * from "./utils";
 export * from "./utils/yagna/yagna";
-export * from "./golem_network";
-export * from "./job";
 export * from "./task";
 export * from "./error/golem-error";
 export { StatsService } from "./stats/service";
